refactor: 이메일 인증 로직 분리
