+++ frontend/src/app/auth/signup/page.tsx
@@ -1,21 +1,18 @@
 'use client';
 
-import { FormField, FormLabel, FormControl, TextField, TextFieldContent, TextFieldButton, Button } from '@wanteddev/wds';
+import { FormField, FormLabel, FormControl, TextField, Button } from '@wanteddev/wds';
 import type { Theme } from '@wanteddev/wds-engine';
 import { useRouter } from 'next/navigation';
 import { useState, useEffect, useMemo } from 'react';
-import { authStorage } from '@/api/authStorage';
+import { EmailVerificationFields } from '@/components/auth/EmailVerificationFields';
 import { usePositionedToast } from '@/components/commons/custom-toast/usePositionedToast';
-import { useCountdown } from '@/hooks/useCountdown';
-import {
-  useSignupMutation,
-  useSendEmailVerificationMutation,
-  useVerifyEmailCodeMutation,
-} from '@/queries/auth';
+import { useAuth } from '@/contexts/AuthContext';
+import { useSignupMutation } from '@/queries/auth';
 
 interface SignupPending {
   socialProvider: string;
   socialAccessToken: string;
+  socialRefreshToken: string;
   name: string;
   email: string;
 }
@@ -34,20 +31,14 @@ const getSignupPending = (): SignupPending | null => {
 
 export default function SignupPage() {
   const router = useRouter();
+  const { login } = useAuth();
   const toast = usePositionedToast();
   const signupMutation = useSignupMutation();
-  const sendEmailMutation = useSendEmailVerificationMutation();
-  const verifyEmailMutation = useVerifyEmailCodeMutation();
 
   const pending = useMemo<SignupPending | null>(() => getSignupPending(), []);
   const [name, setName] = useState(pending?.name || '');
-  const [email, setEmail] = useState(pending?.email || '');
-  const [verificationCode, setVerificationCode] = useState('');
   const [isEmailVerified, setIsEmailVerified] = useState(false);
   const [verifiedEmail, setVerifiedEmail] = useState('');
-  const [isCodeSent, setIsCodeSent] = useState(false);
-  const [verificationError, setVerificationError] = useState(false);
-  const { timeLeft, start: startTimer, reset: resetTimer, formatTime } = useCountdown();
 
   useEffect(() => {
     if (!pending) {
@@ -55,108 +46,24 @@ export default function SignupPage() {
     }
   }, [router, pending]);
 
-  const handleSendVerification = () => {
-    if (!pending) return;
-
-    sendEmailMutation.mutate(
-      { email: email.trim() },
-      {
-        onSuccess: (data) => {
-          if (data.code === 'SEND_EMAIL_VERIFICATION') {
-            setIsCodeSent(true);
-            setIsEmailVerified(false);
-            setVerifiedEmail('');
-            setVerificationError(false);
-            setVerificationCode('');
-            startTimer(300);
-            toast({
-              content: data.message,
-              variant: 'positive',
-              placement: 'top-center',
-            });
-          } else {
-            toast({
-              content: data.message,
-              variant: 'negative',
-              placement: 'top-center',
-            });
-          }
-        },
-        onError: (error) => {
-          toast({
-            content: error.message,
-            variant: 'negative',
-            placement: 'top-center',
-          });
-        },
-      },
-    );
-  };
-
-  const { mutate: verifyEmail, isPending: isVerifying } = verifyEmailMutation;
-
-  useEffect(() => {
-    if (verificationCode.length !== 6) return;
-    if (isEmailVerified || !isCodeSent || timeLeft === 0 || isVerifying) return;
-    if (verificationError) return;
-    if (!pending) return;
-
-    verifyEmail(
-      {
-        email: email.trim(),
-        code: verificationCode.trim(),
-      },
-      {
-        onSuccess: (data) => {
-          if (data.code === 'VERIFY_EMAIL') {
-            setIsEmailVerified(true);
-            setVerifiedEmail(email.trim());
-            setVerificationError(false);
-            toast({
-              content: data.message,
-              variant: 'positive',
-              placement: 'top-center',
-            });
-          } else {
-            setVerificationError(true);
-            setIsEmailVerified(false);
-            toast({
-              content: data.message,
-              variant: 'negative',
-              placement: 'top-center',
-            });
-          }
-        },
-        onError: (error) => {
-          setVerificationError(true);
-          setIsEmailVerified(false);
-          toast({
-            content: error.message,
-            variant: 'negative',
-            placement: 'top-center',
-          });
-        },
-      },
-    );
-  }, [verificationCode, isEmailVerified, isCodeSent, timeLeft, isVerifying, verificationError, pending, email, verifyEmail, toast]);
-
   const handleSubmit = async (e: React.FormEvent<HTMLFormElement>) => {
     e.preventDefault();
-    if (!pending || email.trim() !== verifiedEmail) return;
+    if (!pending || !verifiedEmail) return;
 
     signupMutation.mutate(
       {
         socialProvider: pending.socialProvider,
         socialAccessToken: pending.socialAccessToken,
+        socialRefreshToken: pending.socialRefreshToken,
         nickname: name.trim(),
-        email: email.trim(),
+        email: verifiedEmail,
       },
       {
         onSuccess: (data) => {
           if (data.code === 'SUCCESS' || data.code === 'SIGNUP') {
             if (data.data?.accessToken && data.data?.refreshToken) {
               sessionStorage.removeItem('signup_pending');
-              authStorage.setTokens(data.data.accessToken, data.data.refreshToken);
+              login(data.data.accessToken, data.data.refreshToken);
               router.replace('/projects');
             } else {
               toast({
@@ -245,98 +152,21 @@ export default function SignupPage() {
               </FormControl>
             </FormField>
 
-            <div className="flex flex-col gap-2">
-              <FormField>
-                <FormLabel
-                  htmlFor="email"
-                  variant="label1"
-                  sx={(theme: Theme) => ({
-                    color: theme.semantic.label.neutral,
-                  })}
-                >
-                  이메일
-                </FormLabel>
-                <FormControl>
-                  <TextField
-                    id="email"
-                    value={email}
-                    onChange={(e) => {
-                      setEmail(e.target.value);
-                      setIsEmailVerified(false);
-                      setVerifiedEmail('');
-                      setIsCodeSent(false);
-                      setVerificationCode('');
-                      setVerificationError(false);
-                      resetTimer();
-                    }}
-                    placeholder="이메일을 입력해 주세요."
-                    type="email"
-                    required
-                    width="100%"
-                    trailingButton={
-                      <TextFieldButton
-                        onClick={handleSendVerification}
-                        disabled={sendEmailMutation.isPending || !email.trim()}
-                      >
-                        {isCodeSent ? '재전송' : '인증하기'}
-                      </TextFieldButton>
-                    }
-                  />
-                </FormControl>
-              </FormField>
-
-              <FormField className="gap-2">
-                <FormControl>
-                  <TextField
-                    value={verificationCode}
-                    onChange={(e) => {
-                      const newValue = e.target.value.replace(/\D/g, '').slice(0, 6);
-                      setVerificationCode(newValue);
-                      setVerificationError(false);
-                    }}
-                    placeholder="인증번호를 입력해 주세요."
-                    disabled={isEmailVerified}
-                    invalid={verificationError}
-                    positive={isEmailVerified}
-                    width="100%"
-                    maxLength={6}
-                    trailingContent={
-                      <>
-                        {isCodeSent && timeLeft === 0 && (
-                          <TextFieldContent variant="text" color="semantic.status.negative">
-                            만료됨
-                          </TextFieldContent>
-                        )}
-                      </>
-                    }
-                  />
-                </FormControl>
-                {isCodeSent && timeLeft > 0 && (
-                  <div className={verificationError || isEmailVerified ? "flex items-center justify-between" : "flex justify-end"}>
-                    {verificationError && (
-                      <p className="text-caption-1 text-status-negative">
-                        인증번호가 잘못되었어요.
-                      </p>
-                    )}
-                    {isEmailVerified && (
-                      <p className="text-caption-1 text-label-alternative">인증에 성공했어요.</p>
-                    )}
-                    <p className="text-caption-1 text-label-normal tabular-nums">
-                      유효시간 {formatTime(timeLeft)}
-                    </p>
-                  </div>
-                )}
-              </FormField>
-            </div>
+            <EmailVerificationFields
+              initialEmail={pending?.email}
+              onVerificationChange={(verified, email) => {
+                setIsEmailVerified(verified);
+                setVerifiedEmail(email);
+              }}
+            />
 
             <Button
               type="submit"
               disabled={
                 signupMutation.isPending ||
                 !name.trim() ||
-                !email.trim() ||
                 !isEmailVerified ||
-                email.trim() !== verifiedEmail
+                !verifiedEmail
               }
               variant="solid"
               color="primary"
